+++ lab3/Streams/Streams/EncryptOutputStream.cpp
@@ -7,7 +7,7 @@ CEncryptOutputStream::CEncryptOutputStream(std::unique_ptr<IOutputDataStream>&& 
 	: COutputStreamDecorator(move(stream)), m_enctyptTable(256)
 {
 	std::iota(m_enctyptTable.begin(), m_enctyptTable.end(), 0);
-	std::shuffle(m_enctyptTable.begin(), m_enctyptTable.end(), std::mt19937(key));	
+	std::shuffle(m_enctyptTable.begin(), m_enctyptTable.end(), std::mt19937((unsigned int)key));	
 }
 
 void CEncryptOutputStream::WriteByte(uint8_t data)
@@ -18,7 +18,6 @@ void CEncryptOutputStream::WriteByte(uint8_t data)
 void CEncryptOutputStream::WriteBlock(const void* srcData, std::streamsize size)
 {
 	auto data = static_cast<const uint8_t*>(srcData);
-
 	for (std::streamsize i = 0; i < size; i++)
 	{
 		WriteByte(*data++);
